+++ MyChildhoodNightmare/player.h
@@ -5,7 +5,7 @@ const sf::Vector2f PLAYER_SIZE = { 53, 75 };
 const sf::Vector2f PLAYER_SPAWN_POS = { 350 , 700 };
 const float PLAYER_MOVE_SPEED = 250;
 
-const float PLAYER_START_HEALTH = 100;
+const int PLAYER_START_HEALTH = 100;
 const int PLAYER_START_SHOOTGUN_AMMO = 7;
 const int PLAYER_START_AK_AMMO = 30;
 const int PLAYER_MAX_AMMO = 99;
